# Version 4.0.7
We've got some quality of live features for all of you =)

*__Please Note__*: The Variant Rule "Use DataNinja's Trainer Revamp" has been removed as it's now part of the new Variant Rule "Trainer Advancement Track". If you were using Data's condensed track please update this value accordingly.

## New Features
- Unowns now generate as a random letter when using species-drag-in (#569)
- Trainers' their Dex now checks all enabled Species Compendiums instead of just PTR Species (#573)
  - You can enable / disable alternative species compendiums as a GM in the Compendium Browser -> Settings tab
- Added Throwing Range to the Capability section on the Token Panel (#575)
- Added Move Ranges in their on-hover effect description on the Token Panel (#575)
- Show Type Effectiveness on trainers if they have any changes from Untyped (ergo neutral to all) (#577)
- Added many more filter options to Species in the Compendium Browser (#548)
- Add proper automation for new Trainer Level-up settings

## Bugfixes
- Fixed RollTables now properly referencing their Compendium Versions instead of a local link, therefore not randomly losing their linking (#578)
- Fixed an issue with the TokenImage Rule Element's path resolution (#558)
- Fixed an issue with the Noxial line from the Sage Fan Dex (#568)
- Fixed [CR] feats not properly nesting in their own class (#582)
- Fix an issue with multiple fainted conditions being applied to one mon when their HP reaches 0

## Compendium Changes
- Fixed a typo in Mentoring (it said TM or Tutor moves -> Egg or Tutor moves) (#566)
- Wind Runner's "Gale Speed" has been changed -> "Force of Winds" (#574)

+++ src/module/actor/base.js
@@ -851,10 +851,12 @@ class PTUActor extends Actor {
 
         if (data.system?.health?.value !== undefined) {
             if (data.system.health.value <= 0 && game.settings.get("ptu", "automation.autoFaint")) {
+                if(this.primaryUpdater?.id !== game.user.id) return;
                 const fainted = this.conditions.bySlug("fainted");
                 if (fainted.length === 0) PTUCondition.FromEffects([{ id: "fainted" }]).then(items => this.createEmbeddedDocuments("Item", items));
             }
             else if (data.system.health.value > 0 && game.settings.get("ptu", "automation.autoFaintRecovery")) {
+                if(this.primaryUpdater?.id !== game.user.id) return;
                 const fainted = this.conditions.bySlug("fainted");
                 if (fainted.length > 0) fainted.forEach(f => f.delete());
             }
@@ -1831,7 +1833,10 @@ class PTUActor extends Actor {
                         2: {
                             source: "Level",
                             mode: "add",
-                            value: (this.type === "character" && game.settings.get("ptu", "variant.trainerRevamp") ? (this.system.level.current + this.system.level.current) : this.system.level.current),
+                            value: (
+                                this.type === "character" && ["data-revamp", "short-track"].includes(game.settings.get("ptu", "variant.trainerAdvancement"))
+                                    ? (this.system.level.current + this.system.level.current) 
+                                    : this.system.level.current),
                         },
                         3: {
                             source: "HP Stat",
